html/browsers/history/the-location-interface/location-tojson.html

<!DOCTYPE html>
<title>Location toJSON</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(() => {
  assert_equals(location.toJSON, undefined)
  const prop = Object.getOwnPropertyDescriptor(location, "toJSON")
  assert_false(prop.enumerable)
  assert_false(prop.writable)
  assert_false(prop.configurable)
})
</script>
